Minor code update to filelineiter.

+++ src/util/filelineiter.h
@@ -76,7 +76,7 @@ class FileLineIterator {
   static const int kBufferSize = 5 * 1024;
   char buffer_[kBufferSize];
   int next_position_after_eol_;
-  int buffer_size_;
+  int64 buffer_size_;
   File* file_;
   std::string line_;
 };
